Fix selection of selection tool, improve code.
Our use of wofi was always supposed to support interchangeability (allow for
the use of rofi, dmenu, whatever) and although it isn't able to work for
everything yet due to not generalizing the arguments for such, we can at least
use the environment to determine "what tool is wofi" like we originally
planned. This change does that.

We also improved prompt with the use of the threading macro, where @sogaiu
pointed out it would work well for this case.

FossilOrigin-Name: 966af1b028b8866c93253508ff9fd55167d5855bf94250dfd128217cd5101967