Remove spawn-utils, fix up environment handling.
When running as an executable, `def`'s that are not functions should not be at
the top level. We depended on such for accessing the environment. To fix this
we are now using a `dyn` and a function that acts as it's accessor.

spawn-utils moved to it's own project which was destined to happen to begin
with.

FossilOrigin-Name: bfd9c2255c656d0ed9ff933a52ce46d018b087e012bc566dfc8e74670b4ae65e